@@ -2,29 +2,32 @@
     "name": "laravel/laravel",
     "type": "project",
     "description": "The Laravel Framework.",
-    "keywords": ["framework", "laravel"],
+    "keywords": [
+        "framework",
+        "laravel"
+    ],
     "license": "MIT",
     "require": {
-        "php": "^8.0.2",
-        "bensampo/laravel-enum": "^6.0",
+        "php": "^8.1",
+        "bensampo/laravel-enum": "^6.3",
         "guzzlehttp/guzzle": "^7.2",
-        "inertiajs/inertia-laravel": "^0.5.2",
-        "laravel/framework": "^9.19",
-        "laravel/jetstream": "^2.9",
-        "laravel/sanctum": "^2.14.1",
-        "laravel/tinker": "^2.7",
+        "inertiajs/inertia-laravel": "^0.6",
+        "laravel/framework": "^10.38",
+        "laravel/jetstream": "^2.15",
+        "laravel/sanctum": "^3.2",
+        "laravel/tinker": "^2.8",
         "phpseclib/phpseclib": "~3.0",
-        "symfony/http-client": "^6.1",
-        "tightenco/ziggy": "^1.0",
+        "symfony/http-client": "^6.2",
+        "tightenco/ziggy": "dev-next",
         "xpaw/php-minecraft-query": "^4.0"
     },
     "require-dev": {
         "fakerphp/faker": "^1.9.1",
-        "laravel/sail": "^1.0.1",
+        "laravel/sail": "^1.18",
         "mockery/mockery": "^1.4.4",
         "nunomaduro/collision": "^6.1",
         "phpunit/phpunit": "^9.5.10",
-        "spatie/laravel-ignition": "^1.0"
+        "spatie/laravel-ignition": "^2.0"
     },
     "autoload": {
         "psr-4": {
@@ -63,6 +66,6 @@
         "preferred-install": "dist",
         "sort-packages": true
     },
-    "minimum-stability": "dev",
+    "minimum-stability": "stable",
     "prefer-stable": true
 }
