@@ -25,10 +25,8 @@ class ServerProvisionStatus extends Command
 
     /**
      * Execute the console command.
-     *
-     * @return int
      */
-    public function handle()
+    public function handle(): int
     {
         $servers = Server::all();
         if (empty($servers)) {
@@ -37,8 +35,9 @@ class ServerProvisionStatus extends Command
 
         foreach ($servers as $server) {
             // If linode is online, installing or starting up, continue to next in for loop.
-            if ($server->status == 'online' || $server->status == 'installing' || $server->status == 'startup')
+            if ($server->status == 'online' || $server->status == 'installing' || $server->status == 'startup') {
                 continue;
+            }
 
             // Http client
             $client = HttpClient::createForBaseUri('https://api.linode.com/', [
@@ -46,7 +45,7 @@ class ServerProvisionStatus extends Command
             ]);
 
             // Send request to get linode and its status
-            $response = $client->request('GET', 'https://api.linode.com/v4/linode/instances/' . $server->id);
+            $response = $client->request('GET', 'https://api.linode.com/v4/linode/instances/'.$server->id);
             $data = json_decode($response->getContent());
 
             // Update server status
