@@ -3,20 +3,20 @@
 namespace App\Http\Controllers;
 
 use App\Models\Server;
-use Illuminate\Http\Request;
+use Illuminate\Http\JsonResponse;
 
 class APIController extends Controller
 {
     /**
      * Get all servers from the database and return as JSON
-     *
-     * @return \Illuminate\Http\JsonResponse
      */
-    public function get_servers() {
+    public function get_servers(): JsonResponse
+    {
         return response()->json(Server::all());
     }
 
-    public function get_server_id($servername) {
+    public function get_server_id($servername): JsonResponse
+    {
         return response()->json(Server::where('server_id', '=', $servername)->get()->first());
     }
 }
