@@ -21,17 +21,18 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-Route::middleware('isadminapi')->get('/users', function() {
+Route::middleware('isadminapi')->get('/users', function () {
     return response()->json(User::all());
 })->name('api-get-users');
 
-Route::middleware('isadminapi')->get('/users/{id}', function($id) {
+Route::middleware('isadminapi')->get('/users/{id}', function ($id) {
     return response()->json(User::where('id', '=', $id)->get()->first());
 })->name('api-get-user');
 
-Route::middleware('isadminapi')->post('/users/{id}/update', function(Request $request) {
-    if ($request->name == null || $request->email == null || $request->role == null)
+Route::middleware('isadminapi')->post('/users/{id}/update', function (Request $request) {
+    if ($request->name == null || $request->email == null || $request->role == null) {
         return response()->json(['error' => 'Missing required form fields']);
+    }
 
     $user = User::where('id', '=', $request->id)->get()->first();
     $user->name = $request->name;
@@ -39,6 +40,7 @@ Route::middleware('isadminapi')->post('/users/{id}/update', function(Request $re
     $user->role = $request->role;
     if ($request->password == null) {
         $user->save();
+
         return response()->json('success');
     }
     $user->password = Hash::make($request->password);
@@ -47,18 +49,22 @@ Route::middleware('isadminapi')->post('/users/{id}/update', function(Request $re
     return response()->json('success');
 })->name('api-edit-user');
 
-Route::middleware('isadminapi')->post('/users/create', function(Request $request) {
-    if ($request->name == null)
+Route::middleware('isadminapi')->post('/users/create', function (Request $request) {
+    if ($request->name == null) {
         return response()->json(['error' => 'Missing name form field']);
+    }
 
-    if ($request->email == null)
+    if ($request->email == null) {
         return response()->json(['error' => 'Missing email form field']);
+    }
 
-    if ($request->role == null)
+    if ($request->role == null) {
         return response()->json(['error' => 'Missing role form field']);
+    }
 
-    if ($request->password == null)
+    if ($request->password == null) {
         return response()->json(['error' => 'Missing password form field']);
+    }
 
     $user = new User();
     $user->name = $request->name;
@@ -71,7 +77,7 @@ Route::middleware('isadminapi')->post('/users/create', function(Request $request
     return response()->json('success');
 })->name('api-create-user');
 
-Route::middleware('isadminapi')->post('/users/{id}', function($id) {
+Route::middleware('isadminapi')->post('/users/{id}', function ($id) {
     User::where('id', '=', $id)->get()->first()->delete();
 
     return response()->json('success');
