@@ -1,11 +1,10 @@
 <?php
 
+use App\Models\Server;
 use Illuminate\Foundation\Application;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use Inertia\Inertia;
-use App\Models\ServerProgress;
-use App\Models\Server;
-use Illuminate\Http\Request;
 
 /*
 |--------------------------------------------------------------------------
@@ -28,7 +27,7 @@ Route::get('/', function () {
 });
 
 // Server management routes
-Route::prefix('/server')->group(function() {
+Route::prefix('/server')->group(function () {
     Route::get('/{servername}', [\App\Http\Controllers\LinodeController::class, 'get_server'])->name('get-server');
     Route::get('/provision/{servername}', [\App\Http\Controllers\LinodeController::class, 'create_server'])->name('provision-server');
 });
@@ -51,7 +50,7 @@ Route::middleware([
 Route::middleware([
     'auth:sanctum',
     config('jetstream.auth_session'),
-    'verified', 'isadmin'
+    'verified', 'isadmin',
 ])->group(function () {
     Route::get('/admin/dashboard', function () {
         return Inertia::render('Admin/Dashboard');
