@@ -10,7 +10,7 @@ class BrowserSessionsTest extends TestCase
 {
     use RefreshDatabase;
 
-    public function test_other_browser_sessions_can_be_logged_out()
+    public function test_other_browser_sessions_can_be_logged_out(): void
     {
         $this->actingAs($user = User::factory()->create());
 
