@@ -10,7 +10,7 @@ class TwoFactorAuthenticationSettingsTest extends TestCase
 {
     use RefreshDatabase;
 
-    public function test_two_factor_authentication_can_be_enabled()
+    public function test_two_factor_authentication_can_be_enabled(): void
     {
         $this->actingAs($user = User::factory()->create());
 
@@ -22,7 +22,7 @@ class TwoFactorAuthenticationSettingsTest extends TestCase
         $this->assertCount(8, $user->fresh()->recoveryCodes());
     }
 
-    public function test_recovery_codes_can_be_regenerated()
+    public function test_recovery_codes_can_be_regenerated(): void
     {
         $this->actingAs($user = User::factory()->create());
 
@@ -39,7 +39,7 @@ class TwoFactorAuthenticationSettingsTest extends TestCase
         $this->assertCount(8, array_diff($user->recoveryCodes(), $user->fresh()->recoveryCodes()));
     }
 
-    public function test_two_factor_authentication_can_be_disabled()
+    public function test_two_factor_authentication_can_be_disabled(): void
     {
         $this->actingAs($user = User::factory()->create());
 
