@@ -11,7 +11,7 @@ class UpdatePasswordTest extends TestCase
 {
     use RefreshDatabase;
 
-    public function test_password_can_be_updated()
+    public function test_password_can_be_updated(): void
     {
         $this->actingAs($user = User::factory()->create());
 
@@ -24,7 +24,7 @@ class UpdatePasswordTest extends TestCase
         $this->assertTrue(Hash::check('new-password', $user->fresh()->password));
     }
 
-    public function test_current_password_must_be_correct()
+    public function test_current_password_must_be_correct(): void
     {
         $this->actingAs($user = User::factory()->create());
 
@@ -39,7 +39,7 @@ class UpdatePasswordTest extends TestCase
         $this->assertTrue(Hash::check('password', $user->fresh()->password));
     }
 
-    public function test_new_passwords_must_match()
+    public function test_new_passwords_must_match(): void
     {
         $this->actingAs($user = User::factory()->create());
 
